changes to challenge participant

+++ community/src/main/java/com/app/community/business/service/impl/ChallengeParticipantServiceImpl.java
@@ -1,9 +1,12 @@
 package com.app.community.business.service.impl;
 
+import com.app.community.auth.util.JwtTokenUtil;
 import com.app.community.business.mapper.ChallengeParticipantMapper;
 import com.app.community.business.repository.ChallengeParticipantRepository;
 import com.app.community.business.repository.model.ChallengeParticipantDAO;
 import com.app.community.business.service.ChallengeParticipantService;
+import com.app.community.dto.ChallengeParticipantUpdateDTO;
+import com.app.community.exception.UnauthorizedException;
 import com.app.community.model.ChallengeParticipant;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -17,23 +20,20 @@ public class ChallengeParticipantServiceImpl implements ChallengeParticipantServ
 
     private final ChallengeParticipantRepository participantRepository;
     private final ChallengeParticipantMapper participantMapper;
-    private final UserServiceClient userServiceClient;
+    private final JwtTokenUtil jwtTokenUtil;
 
     public ChallengeParticipantServiceImpl(ChallengeParticipantRepository participantRepository,
                                            ChallengeParticipantMapper participantMapper,
-                                           UserServiceClient userServiceClient) {
+                                           JwtTokenUtil jwtTokenUtil) {
         this.participantRepository = participantRepository;
         this.participantMapper = participantMapper;
-        this.userServiceClient = userServiceClient;
+        this.jwtTokenUtil = jwtTokenUtil;
     }
 
     @Override
-    public ChallengeParticipant joinChallenge(Long challengeId, Long userId) {
+    public ChallengeParticipant joinChallenge(Long challengeId, String token) {
         log.info("Checking if challenge exists");
-
-        if (!userServiceClient.doesUserExist(userId)) {
-            throw new IllegalArgumentException("User with ID " + userId + " does not exist");
-        }
+        Long userId = jwtTokenUtil.extractUserId(token.replace("Bearer ", ""));
 
         List<ChallengeParticipantDAO> existingParticipants = participantRepository.findByUserIdAndChallengeId(userId, challengeId);
         if (!existingParticipants.isEmpty()) {
@@ -42,9 +42,9 @@ public class ChallengeParticipantServiceImpl implements ChallengeParticipantServ
 
         log.info("Joining a challenge, user and challenge ids : {}, {}", userId, challengeId );
         ChallengeParticipantDAO participantDAO = new ChallengeParticipantDAO();
-        participantDAO.setUser(userId);
-        participantDAO.setChallenge(challengeId);
-        log.info("saving to db, user and challenge ids : {}, {}", participantDAO.getUser(), participantDAO.getChallenge() );
+        participantDAO.setUserId(userId);
+        participantDAO.setChallengeId(challengeId);
+        log.info("saving to db, user and challenge ids : {}, {}", participantDAO.getUserId(), participantDAO.getChallengeId() );
         participantDAO.setProgress(0);
         participantDAO.setCompleted(false);
         participantDAO.setJoinedAt(java.time.LocalDateTime.now());
@@ -63,13 +63,22 @@ public class ChallengeParticipantServiceImpl implements ChallengeParticipantServ
 
     @Transactional
     @Override
-    public ChallengeParticipant updateProgress(Long challengeId, Long userId, int progress) {
+    public ChallengeParticipant updateProgress(Long challengeId, ChallengeParticipantUpdateDTO challengeParticipantUpdateDTO,
+                                               String token) {
         log.info("Updating progress");
+        Long userId = jwtTokenUtil.extractUserId(token.replace("Bearer ", ""));
+        log.info("user from token " + userId);
+        log.info("challenge id " + challengeId);
         List<ChallengeParticipantDAO> participants = participantRepository.findByUserIdAndChallengeId(userId, challengeId);
         if (participants.isEmpty()) {
+            log.info("User is not a participant of this challenge");
             throw new IllegalStateException("User is not a participant of this challenge");
         }
-
+        if (!isAuthorized(token, participants.get(0).getUserId())) {
+            log.warn("Unauthorized attempt to update review ID {} by user ID {}", challengeId, userId);
+            throw new UnauthorizedException("You are not authorized to update this review");
+        }
+        Integer progress = challengeParticipantUpdateDTO.getProgress();
         ChallengeParticipantDAO participantDAO = participants.get(0);
         participantDAO.setProgress(progress);
         participantDAO.setCompleted(progress == 100);
@@ -77,12 +86,26 @@ public class ChallengeParticipantServiceImpl implements ChallengeParticipantServ
     }
 
     @Override
-    public void leaveChallenge(Long challengeId, Long userId) {
+    public void leaveChallenge(Long challengeId, String token) {
+        Long userId = jwtTokenUtil.extractUserId(token.replace("Bearer ", ""));
         List<ChallengeParticipantDAO> participants = participantRepository.findByUserIdAndChallengeId(userId, challengeId);
+        if (!isAuthorized(token, participants.get(0).getUserId())) {
+            log.warn("Unauthorized attempt to update review ID {} by user ID {}", challengeId, userId);
+            throw new UnauthorizedException("You are not authorized to update this review");
+        }
+
         if (participants.isEmpty()) {
             throw new IllegalStateException("User is not a participant of this challenge");
         }
+
         log.info("Leaving a challenge");
         participantRepository.delete(participants.get(0));
+    }
+
+    private boolean isAuthorized(String token, Long userId) {
+        String cleanToken = token.replace("Bearer ", "");
+        Long tokenUserId = jwtTokenUtil.extractUserId(cleanToken);
+        String role = jwtTokenUtil.extractRole(cleanToken);
+        return tokenUserId.equals(userId) || "ADMIN".equals(role);
     }
 }
